==> 01-eda.ipynb <==
# EDA for Flavor Network Data

import pandas as pd
import numpy as np

# This is our regional recipe component data
dishes = pd.read_csv('Data/dish_components.csv',error_bad_lines=False)

# Transpose makes dishes easier to index and we drop a few empty rows
dishes = dishes.T
dishes = dishes.dropna(how='all')

# This is our shared compound data
compounds = pd.read_csv('Data/shared_compounds.csv')
pd.DataFrame(compounds)



# This is the GroupBy object that allows lookup of number of shared compounds
combinator = compounds.copy()
combinatrix = combinator.groupby(['First Ingredient','Second Ingredient']).sum()
combinatrix

# Make a zeros matrix with correct labels

# Make a list of all unique ingredients.
Ingredients = set(compounds['First Ingredient'])
len(Ingredients)
Ingredients2 = set(compounds['Second Ingredient'])
Ingredients.update(Ingredients2)
len(Ingredients)
unique = sorted(list(Ingredients))
dimen = len(unique)

# Zeros matrix with columns and index of unique ingredients
flavor_mat = np.ma.zeros((dimen,dimen), dtype='Int16')
flavor_matrix = pd.DataFrame(flavor_mat, index=unique, columns=unique)
flavor_matrix.head(25)

# Make 2 matrices: shared compounds, and ingredients in the same recipe

## Shared compounds matrix:

# List shared flavor compounds by ingredient. Remove 'int' to double-check locations.
long_list = []

for i in range(0,len(unique)):
    for j in range(0,len(unique)):
        try: 
            long_list.append(int(combinatrix.loc[unique[i], unique[j]]))
        except KeyError:
            long_list.append(0)

# Make flavor compounds list an array
long_list_array = np.array(long_list).reshape((len(unique),len(unique)))

# Put array into dataframe with column and index labels
comp_share = pd.DataFrame(long_list_array, index=unique, columns=unique)

# Make the matrix symmetrical here by adding the transposed matrix
comp_share = comp_share + comp_share.T

# Save to csv
#comp_share.to_csv('Shared_Compounds_Dense.csv')

comp_share = pd.read_csv('Shared_Compounds_Dense.csv')

## Concurrent ingredients from recipes matrix

# Define function for getting a list of components in recipes. Recipes assumed to be in columns here.
def get_rec_compon(recipes):
    rec_compon = []
    for i in range(0,len(recipes.columns)):
        irec = recipes[i]
        irec = irec.dropna()
        irec = irec[1:][0:]
        icomp = [component for component in irec]
        rec_compon.append(icomp)
    return rec_compon        

# This is a list of components
dish_compon = get_rec_compon(dishes)

# Copy the zeros matrix to make sure we're starting from scratch
rec_compon_matrix = flavor_matrix.copy()

# Now we can loop through our list of dish components.
for components in dish_compon:
    for i in range(0,len(components)):
        for j in range(0,len(components)):
            if components[i] != components[j]:
                try:
                    rec_compon_matrix.loc[components[i]][components[j]] += 1
                except KeyError:
                    #print("Couldn't find ", str(components[i]), "/", str(components[j]))
                    continue
            else:
                continue

# Note on above cell: 'starch' occurs in many recipes, but shares no flavor compounds with anything else
# therefore it is not in the list of unique ingredients. There are probably others like this.

# Save recipe components matrix
#rec_compon_matrix.to_csv("Recipe_Components_Dense.csv")

rec_compon_matrix = pd.read_csv('Recipe_Components_Dense.csv')

# Drill down into specific recipes and look at the flavor compound relationships, then visualize with a Chort chart

# Take a column from dishes, drop nan and region of origin
recipe = dishes[1104]
recipe = recipe.dropna()
recipe = recipe[1:][0:]
recipe

# List comprehension for recipe
compon = [ing for ing in recipe]

# Testing loc on GroupBy object
combinatrix.loc[compon[3],compon[1]]

# List shared flavor compounds in recipe. Remove 'int' to check locations.
components = []

for i in range(0,len(compon)):
    for j in range(0,len(compon)):
        try: 
            components.append(int(combinatrix.loc[compon[i], compon[j]]))
        except KeyError:
            components.append(0)

# Make list into array
dish_shared = np.array(components).reshape((len(recipe),len(recipe)))

# Creates DataFrame and Matrix as a list.
cereals = pd.DataFrame(dish_shared, index=compon, columns=compon)

# Make the matrix symmetrical here by adding the transposed matrix
cereals = cereals + cereals.T

hops = cereals.values.tolist()

# Values for Chord chart
hops

# Labels for Chord chart
compon

from chord import Chord

#First Chord chart
#Chord(hops, compon).to_html('chord.html')
