==> 02-chords_and_flavors.ipynb <==
# Chord charts and the Flavor Generator

import pandas as pd
import numpy as np
from chord import Chord
import random

# Load and set up data
dishes = pd.read_csv('Data/dish_components.csv',error_bad_lines=False)
numdish = len(dishes)
dishes = dishes.T
dishes = dishes.dropna(how='all')
compounds = pd.read_csv('Data/shared_compounds.csv')
combinatrix = compounds.groupby(['First Ingredient','Second Ingredient']).sum()

# Random dish to flavor Chord chart

# Pick out a random dish
rand_dish = int(np.random.randint(numdish, size=1))

# Look up the recipe, get the ingredients
recipe = dishes[rand_dish]
recipe = recipe.dropna()
recipe = recipe[1:][0:]
compon = [ing for ing in recipe]
recipe

# Loop through the ingredients and record values for shared flavor compounds
components = []

for i in range(0,len(compon)):
    for j in range(0,len(compon)):
        try: 
            components.append(int(combinatrix.loc[compon[i], compon[j]]))
        except KeyError:
            components.append(0)
    

# Make list into array
component_array = np.array(components).reshape((len(recipe),len(recipe)))

# Creates DataFrame and Matrix as a list.
cereals = pd.DataFrame(component_array, index=compon, columns=compon)

# Make symmetrical for chord chart
cereals = cereals + cereals.T

hops = cereals.values.tolist()

cereals

compon

# Random recipe Chord chart
# Chord(hops, compon).to_html('Dish' + str(rand_dish)+ '.html')

# Novel flavor combination generation

# Load in matrices from EDA.
shares = pd.read_csv("Shared_Compounds_Dense.csv")
recs = pd.read_csv("Recipe_Components_Dense.csv")

# Flavor generator function gives a list of flavors from a starting flavor
def flavor_generator(first_flavor, num_flavors, flavor_space):
    flavors = []
    flavors.append(first_flavor)
    n = num_flavors - 1
    current_flavor = first_flavor
    for i in range(0, n):
        new_flavor_index = random.choices(flavor_space.index,
                                          weights=flavor_space[current_flavor], k=1)
        new_flavor = flavor_space.iloc[new_flavor_index[0]][0]
        flavors.append(new_flavor)
        current_flavor = new_flavor
    return flavors

# Flavor dataframe function shows the flavor connections between ingredients in a list
def flavor_df(flavors):
    components = []

    for i in range(0,len(flavors)):
        for j in range(0,len(flavors)):
            try: 
                components.append(int(combinatrix.loc[flavors[i], flavors[j]]))
            except KeyError:
                components.append(0)
                
    flav_ar = np.array(components).reshape((len(flavors),len(flavors)))
    flav_df = pd.DataFrame(flav_ar, index=flavors, columns=flavors)
    flav_df = flav_df + flav_df.T
    
    return flav_df

# This is a result when applying the flavor generator to the recipes matrix
flavor_generator('coriander',7,recs)

# This is a result when applying the flavor generator to the shared compounds matrix
flavor_generator('enokidake',7,shares)

# Generate a few lists to use with the flavor dataframe function
flavor_6 = flavor_generator('cayenne',6,shares)
flavor_7 = flavor_generator('coriander',7,shares)
flavor_8 = flavor_generator('matsutake',7,shares)

fen = ['fennel','enokidake','coriander','marjoram','celery']
fenn = ['fennel', 'enokidake', 'feta_cheese', 'cherimoya', 'jamaican_rum']

# This is the cayenne dataframe
flavor_df(fenn)

# This is the coriander dataframe
flavor_df(flavor_7)

# This is the matsutake dataframe
flavor_df(flavor_8)

# We can just take the values for our chord chart
lister = flavor_df(flavor_6).values.tolist()
lister

# The labels for the chord chart is the input for the flavor dataframe function
#Chord(lister, flavor_7).to_html('cayenne_6.html')
