==> 04-cluster_matrix_construction.ipynb <==
# Condense recipes into cluster matrices

import pandas as pd
import numpy as np
import re
import string

# Load TFIDF file and cluster labels
data = pd.read_csv('rec_tfidf.csv')
labels = pd.read_csv('TFIDF_Cluster_Labels.csv')

# Assign cluster labels
labels.rename(columns={'0':'Labels'})
data = data.rename(columns={'Unnamed: 0':'Cluster'})
data['Cluster'] = labels['0']

# Groupby object for cluster centroids
gb_cluster = data.groupby('Cluster').mean()

#gb_cluster.to_csv('cluster_lookup.csv')

# Boolean for cluster ingredients
clustermask = gb_cluster != 0.0
clustermask

# Transposed matrix for functions
clusterind = clustermask.T
clusterind

# Define a cluster ingredient list, then a recipe ingredient list. 
def cluster_ing(cluster_num):
    all_ing = (clustermask.iloc[cluster_num] == True).index
    cluster_ing = []
    for i in all_ing:
        if clusterind[cluster_num][i] == True:
            cluster_ing.append(i)
        else:
            continue
    return cluster_ing

# List of ingredients for smallest cluster
cluster_ing(1)

## Locate recipes in data

# Fill the zeros matrix with recipe components
    for components in dish_compon:
    for i in range(0,len(components)):
        for j in range(0,len(components)):
            if components[i] != components[j]:
                try:
                    rec_compon_matrix.loc[components[i]][components[j]] += 1
                except KeyError:
                    #print("Couldn't find ", str(components[i]), "/", str(components[j]))
                    continue
            else:
                continue

# Index object
clust_ing = (clustermask.iloc[1] == True).index
# Empty set
cluster_ing = []
# For loop for index object
for i in clust_ing:
    # Running through groupby object row
    if clusterind[1][i] == True:
        # Append to empty set
        cluster_ing.append(i)
    else:
        continue
cluster_ing

databools = data != 0
databools['Cluster'] = data['Cluster']

recipes_1.index

## Main recipe matrix function

# Define a function that outputs a recipe matrix
def recipe_matrix(cluster_num):
    
    # Define the ingredients list for the cluster
    # Index object of possible ingredients
    clust_ing = (clustermask.iloc[cluster_num] == True).index
    # Empty list
    cluster_ing = []
    # Loop through possible ingredients in index object
    for i in clust_ing:
        # Look in groupby object (row where True means that the ingredient is in that cluster)
        if clusterind[cluster_num][i] == True:
            # Append index (ingredient) to empty list
            cluster_ing.append(i)
        else:
            continue

    # Define the recipes matrix
    cluster_recs = pd.DataFrame(databools.loc[databools.Cluster == cluster_num], columns=cluster_ing)
        
    # Define which indices (recipes) are in the cluster
    indices = cluster_recs.index
    
    # Empty list for cluster recipe ingredients
    cluster_rec_ings = []
    
    for ind in indices:
        # Define the recipe
        rec = cluster_recs.loc[ind]
        # Empty list for recipe ingredients        
        rec_ing = []
        # For loop of ingredients in the cluster
        for i in cluster_ing:
        # Conditional for this particular recipe
            if rec[i] == True:
                # Add ingredient to recipe ingredients
                rec_ing.append(i)
            else:
                continue
        # Add recipe to cluster recipes        
        cluster_rec_ings.append(rec_ing)
    
    # Define the zeros matrix
    flavor_mat = np.ma.zeros((len(cluster_ing),len(cluster_ing)), dtype='Int16')
    flavor_matrix = pd.DataFrame(flavor_mat, index=cluster_ing, columns=cluster_ing)
    
    # Loop through the cluster recipe ingredients and add 1 to all the places where ingredients occur
    for rec in cluster_rec_ings:
        if rec != []:
            first_ingredient = rec[0]
            rec.remove(first_ingredient)
            if rec != []:
                for other_ingredient in rec:
                    flavor_matrix[first_ingredient].loc[other_ingredient] += 1
            else:
                continue
        else:
            continue
    
    # Add the transposed matrix so that the output is symmetrical
    cluster_matrix = flavor_matrix + flavor_matrix.T
        
    return cluster_matrix

# Clusters without matrices
clust_num = [4,5,6,7,8,9,10,11,12,13]

# Loop to apply function to make additional matrices
for clust in clust_num:
    matrix = recipe_matrix(clust)
    matrix.to_csv('cluster_'+str(clust)+'_recipe_matrix.csv')

## Never use list comprehension for recipe matrices

## Stop, this is very slow
# [ing for ing in databools.iloc[0].index if databools.iloc[0][ing] == True]

## Visualize ingredient concurrence to make sure that this works as intended

guinea_pig = cluster_rec_ings[1].copy()
gp_mat = np.ma.zeros((len(guinea_pig),len(guinea_pig)), dtype='Int16')
gp_matrix = pd.DataFrame(gp_mat, index=guinea_pig, columns=guinea_pig)

# Try with one recipe first.
while guinea_pig != []:
    pig = guinea_pig[0]
    guinea_pig.remove(pig)
    if guinea_pig != []:
        for ing in guinea_pig:
            gp_matrix[pig].loc[ing] += 1
    else:
        continue

gp_matrix
