==> 05-cluster_explorer.ipynb <==
# Get a flavor pairing from the flavor generator, find the nearest cluster, pull the rest of the ingredients from that cluster

import pandas as pd
import numpy as np
from sklearn.metrics import pairwise
import random

# Define flavor matrices here
recs = pd.read_csv("Recipe_Components_Dense.csv")
shares = pd.read_csv("Shared_Compounds_Dense.csv")

# Import cluster centroids
cl_gb = pd.read_csv('Clusters/cluster_lookup.csv')

# Let's not keep the target column (Cluster)
cl_gb = cl_gb.drop(columns='Cluster')

## Make a novel flavor combination

def flavor_generator(first_flavor, num_flavors, flavor_space):
    flavors = []
    flavors.append(first_flavor)
    n = num_flavors - 1
    current_flavor = first_flavor
    fail = 0
    while n > 0:
        new_flavor_index = random.choices(flavor_space.index,
                                          weights=flavor_space[current_flavor], k=1)
        new_flavor = flavor_space.iloc[new_flavor_index[0]][0]
        if new_flavor in flavors:
            fail += 1
            if fail > 10000:
                print("Too much failure. Can't continue.")
                break
            else:
                current_flavor = first_flavor
                continue
        else:
            flavors.append(new_flavor)
            print('Went from ', current_flavor, ' to ', new_flavor, '. ', 
                  flavor_space[current_flavor].iloc[new_flavor_index[0]], ' shared flavor compounds.')
            current_flavor = new_flavor
            n -= 1
    return flavors

gen_1 = flavor_generator('bacon', 5, shares)
gen_1

## Find a cluster for the flavor combination

def cluster_choice(flavor_set):
    gen_rec = np.ma.zeros(len(cl_gb.columns), dtype='Int16')
    gen_ing = pd.Series(gen_rec, index=cl_gb.columns)
    gen_2 = []
    for ing in flavor_set:
        twos = ing.split('_')
        gen_2.extend(twos)
        
    for ing in gen_2:
        try:
            gen_ing[ing] += 1
        finally:
            continue
    
    cosinesim = pairwise.cosine_similarity(np.array(gen_ing).reshape(1, -1), cl_gb)
    cosinesim = cosinesim[0]
    clusterchoix = np.argmax(cosinesim)
    return flavor_set, clusterchoix, cosinesim

cl_map = cluster_choice(gen_1)
cl_map

# Cluster can be changed here, cosine distance visible
cl_map_revised = [cl_map[0], 3, cl_map[2]]

## Bring in additional flavors from that cluster

def cluster_explorer(cluster_map, num_ingredients):
    ingredients = []
    flavor_space = pd.read_csv('Clusters/cluster_' + str(cluster_map[1]) + '_recipe_matrix.csv')
    gen_2 = []
    
    for ing in cluster_map[0]:
        twos = ing.split('_')
        gen_2.extend(twos)
    
    liftoff = random.choice(gen_2)
    jumpoff = liftoff
    failures = 0
    
    while len(set(ingredients)) < num_ingredients:
        if failures > 500:
            print('Failed to find ingredients in cluster ', str(cluster_map[1]), '!')
            break
        else:
            try:
                new_ind = random.choices(flavor_space.index,
                                         weights=np.log(flavor_space[liftoff] + 1), k=1)
                new_flavor = flavor_space.iloc[new_ind[0]][0]
                if new_flavor in ingredients:
                    liftoff = random.choice(gen_2)
                    failures += 1
                    continue
                else:
                    ingredients.append(new_flavor)
                    print(liftoff, ' and ', new_flavor, ' appear together in ', 
                          flavor_space[liftoff].iloc[new_ind[0]], ' recipes in cluster ', cluster_map[1] )
                    #liftoff = new_flavor
            except KeyError:
                liftoff = random.choice(gen_2)
                failures += 1
                continue
    print('Started with', jumpoff, '...')    
    return cluster_map[0], set(ingredients)

cluster_explorer(cl_map_revised, 8)
